Remove duplicate Sass lint task.
