+++ gulp-tasks/lint.js
@@ -1,21 +1,45 @@
 const { src } = require('gulp');
+const { formatters } = require("stylelint");
 const gulpStylelint = require('gulp-stylelint');
 const eslint = require('gulp-eslint');
+const dutil = require('./utils/doc-util');
+
+const IGNORE_STRING = "This file is ignored";
+
+function ignoreStylelintIgnoreWarnings(lintResults) {
+  return formatters.string(
+    lintResults.reduce((memo, result) => {
+      const { warnings } = result;
+      const fileIsIgnored = warnings.some((warning) =>
+        RegExp(IGNORE_STRING, "i").test(warning.text)
+      );
+
+      if (!fileIsIgnored) {
+        memo.push(result);
+      }
+
+      return memo;
+    }, [])
+  );
+}
 
 module.exports = {
-  // Lint Sass based on .stylelintrc.yml config.
+  // Lint Sass based on .stylelintrc.json config.
   lintSass () {
-    return src(['./src/patterns/stylesheets/**/*.scss'])
-      .pipe(
-        gulpStylelint({
-          reporters: [
-            {
-              formatter: 'string',
-              console: true
-            }
-          ]
-        })
-      );
+    return src("./src/patterns/stylesheets/**/*.scss")
+    .pipe(
+      gulpStylelint({
+        failAfterError: true,
+        reporters: [
+          {
+            formatter: ignoreStylelintIgnoreWarnings,
+            console: true,
+          },
+        ],
+        syntax: "scss",
+      })
+    )
+    .on("error", dutil.logError);
   },
 
   // Lint JavaScript based on .eslintrc config.
